Update the auto-testing to get Travis CI working

+++ src/PyRsw.py
@@ -3,9 +3,10 @@
 ##
 
 import numpy as np
-import matplotlib.pyplot as plt
+import matplotlib
 import Plot_tools
 import Diagnose
+import Steppers
 from scipy.fftpack import fftn, ifftn, fftfreq
 import os, sys, shutil
 
@@ -52,9 +53,11 @@ class Simulation:
         self.g    = 9.81            # gravity
         self.f0   = 1e-4            # Coriolis
         self.beta = 0.
-        self.cfl  = 0.5             # default CFL
-        self.time = 0               # initial time
+        self.cfl  = -1             # default CFL
+        self.time = 0.              # initial time
         self.min_dt = 1e-3          # minimum timestep
+        self.adaptive = True        # Adaptive (True) or fixed (False) timestep
+        self.fixed_dt = 1.
 
         self.geomx = 'periodic'     # x boundary condition
         self.geomy = 'periodic'     # y boundary condition
@@ -75,14 +78,27 @@ class Simulation:
 
         self.plott = np.inf
         self.diagt = np.inf
+        self.savet = np.inf
 
         self.num_steps = 0
         self.mean_dt = 0.
+
+        self.restarting = False
+        self.restart_index = 0
 
         self.topo_func = null_topo  # Default to no topograpy
         
     # Full initialization for once the user has specified parameters
     def initialize(self):
+
+        # Determine the CFL value based on the time-stepping scheme
+        if self.cfl == -1:
+            if self.stepper.__name__ == 'AB3':
+                self.cfl = 0.5
+            elif self.stepper.__name__ == 'AB2':
+                self.cfl = 0.05 # Needs to be confirmed!!
+            elif self.stepper.__name__ == 'Euler':
+                self.cfl = 0.005 # Needs to be confirmed!!
 
         print('Parameters:')
         print('-----------')
@@ -91,6 +107,7 @@ class Simulation:
         if self.Ny > 1:
             print('geomy    = {0:s}'.format(self.geomy))
         print('stepper  = {0:s}'.format(self.stepper.__name__))
+        print('CFL      = {0:g}'.format(self.cfl))
         print('method   = {0:s}'.format(self.method))
         print('dynamics = {0:s}'.format(self.dynamics))
         print('Nx       = {0:d}'.format(self.Nx))
@@ -101,6 +118,8 @@ class Simulation:
                 print('Coriolis = beta-plane')
             else:
                 print('Coriolis = f-plane')
+        if self.restarting:
+            print('Restarting from index {0:d}'.format(self.restart_index))
         print ' '
         
         # Initialize grids and cell centres
@@ -169,7 +188,34 @@ class Simulation:
             filty = np.array([1.0])
                 
         self.sfilt = np.tile(filtx,(1,self.Nky))*np.tile(filty,(self.Nkx,1))
+
+        # If we're using a fixed dt, check that it matches plott, savet, and diagt
+        if self.animate.lower() != 'none':
+            if self.plott/self.fixed_dt != int(self.plott/self.fixed_dt):
+                print('Error: Plot interval not integer multiple of fixed dt.')
+                sys.exit()
+        if self.diagnose:
+            if self.diagt/self.fixed_dt != int(self.diagt/self.fixed_dt):
+                print('Error: Diagnostic interval not integer multiple of fixed dt.')
+                sys.exit()
+        if self.output:
+            if self.savet/self.fixed_dt != int(self.savet/self.fixed_dt):
+                print('Error: Output interval not integer multiple of fixed dt.')
+                sys.exit()
+
             
+        # If we're restarting load the appropriate files
+        if self.restarting:
+            try:
+                restart_file = np.load('Outputs/{0:s}/{1:05d}.npz'.format(self.run_name,self.restart_index))
+                self.soln.u = restart_file['u']
+                self.soln.v = restart_file['v']
+                self.soln.h = restart_file['h']
+                self.time = restart_file['t']
+            except:
+                print('Restarting failed. Exiting.')
+                sys.exit()
+
         
     def prepare_for_run(self):
 
@@ -178,7 +224,8 @@ class Simulation:
     
         # If we're saving, initialize those too
         if self.output:
-            self.next_save_time = self.savet
+            self.next_save_time = (self.restart_index+1)*self.savet
+            self.out_counter = self.restart_index
     
         # If we're saving, initialize the directory
         if self.output or (self.animate == 'Save') or self.diagnose:
@@ -203,8 +250,10 @@ class Simulation:
             self.hov_count = 1
 
             self.initialize_plots(self)
-            self.update_plots(self)
-            self.next_plot_time = self.plott
+            if not(self.restarting):
+                self.update_plots(self)
+            self.frame_count = int(np.floor(self.time/self.plott) + 1)
+            self.next_plot_time = self.frame_count*self.plott
 
     # Compute the current flux
     def flux(self):
@@ -215,7 +264,7 @@ class Simulation:
         
         t = self.time + self.dt
 
-        nt = np.Inf # next time for doing stuff
+        nt = self.end_time # next time for doing stuff
         if self.animate != 'None':
             nt = min([self.next_plot_time, nt])
         if self.diagnose:
@@ -223,7 +272,7 @@ class Simulation:
         if self.output:
             nt = min([self.next_save_time, nt])
 
-        if nt < t:
+        if (nt < t) and self.adaptive:
             self.dt = nt - self.time
 
         t = self.time + self.dt
@@ -248,8 +297,9 @@ class Simulation:
     # Advance the simulation one time-step.
     def step(self):
 
+        #FJP: comment this back in
         self.compute_dt() 
-
+        
         # Check if we need to adjust the time-step
         # to match an output time
         do_plot, do_diag, do_save = self.adjust_dt()
@@ -296,15 +346,19 @@ class Simulation:
             if self.Nz == 1:
                 pstr  = '{0:s}'.format(Plot_tools.smart_time(self.time))
                 pstr += ' '*(13-len(pstr))
-                pstr += ', dt = {0:0<7.1e}'.format(self.mean_dt)
-                L = len(pstr) - 11
+                try:
+                    pstr += ',  dt = {0:0<7.1e}'.format(mean(self.dts))
+                except:
+                    pstr += ',  dt = {0:0<7.1e}'.format(self.dt)
+                L = len(pstr) - 13
                 pstr += ', min(u,v,h) = ({0: < 8.4e},{1: < 8.4e},{2: < 8.4e})'.format(minu,minv,minh)
                 pstr += ', del_mass = {0: .2g}'.format(mass/self.Ms[0]-1)
                 pstr += '\n'
                 tmp = '  = {0:.3%}'.format(self.time/self.end_time)
                 pstr += tmp
-                pstr += ' '*(L + 13 - len(tmp))            
-                pstr += 'max(u,v,h) = ({0: < 8.4e},{1: < 8.4e},{2: < 8.4e})'.format(maxu,maxv,maxh)
+                pstr += ' '*(L - len(tmp))            
+                pstr += 'avg = {0:0<7.1e}'.format(self.mean_dt)
+                pstr += ', max(u,v,h) = ({0: < 8.4e},{1: < 8.4e},{2: < 8.4e})'.format(maxu,maxv,maxh)
                 pstr += ', del_enrg = {0: .2g}'.format(enrg/(self.KEs[0]+self.PEs[0])-1)
             else:
                 pstr  = 't = {0:.4g}s'.format(self.time)
@@ -316,7 +370,15 @@ class Simulation:
                 pstr += ', del_enrg = {0:+.2g}'.format(enrg/(self.KEs[0]+self.PEs[0])-1)
                 pstr += '\n'
                 pstr += '  = {0:.3%}'.format(self.time/self.end_time)
-            print('\n{0:s}: output {1:d} : step {2:d}'.format(self.run_name, int(self.time/self.plott),self.num_steps))
+
+            head_str = ('\n{0:s}'.format(self.run_name))
+            if self.animate != 'None':
+                head_str += ': frame {0:d}'.format(self.frame_count-1)
+            if self.output:
+                head_str += ': output {0:d}'.format(self.out_counter-1)
+            head_str += ': step {0:d}'.format(self.num_steps)
+
+            print(head_str)
             print(pstr)
 
     # Step until end-time achieved.
@@ -331,39 +393,57 @@ class Simulation:
             Diagnose.plot(self)
         
         if (self.animate == 'Anim'):
-            plt.ioff()
-            plt.show()
+            matplotlib.pyplot.ioff()
+            matploblib.show()
 
         if self.diagnose:
             Diagnose.save(self)
 
     # Compute time-step using CFL condition.
     def compute_dt(self):
-        c = np.sqrt(self.g*np.sum(self.Hs))
-        eps = 1e-8
 
-        max_u = np.max(np.abs(self.soln.u.ravel()))
-        max_v = np.max(np.abs(self.soln.v.ravel()))
+        if self.adaptive:
+            c = np.sqrt(self.g*np.sum(self.Hs))
+            eps = 1e-8
 
-        dt_x = self.end_time - ((self.Nx-1)/(self.Nx-1+eps))*(self.end_time - self.dx[0]/(max_u+2*c))
-        dt_y = self.end_time - ((self.Ny-1)/(self.Ny-1+eps))*(self.end_time - self.dx[1]/(max_v+2*c))
+            max_u = np.max(np.abs(self.soln.u.ravel()))
+            max_v = np.max(np.abs(self.soln.v.ravel()))
 
-        self.dt = max([self.cfl*min([dt_x,dt_y]),self.min_dt])
+            dt_x = self.end_time - ((self.Nx-1)/(self.Nx-1+eps))*(self.end_time - self.dx[0]/(max_u+2*c))
+            dt_y = self.end_time - ((self.Ny-1)/(self.Ny-1+eps))*(self.end_time - self.dx[1]/(max_v+2*c))
+
+            self.dt = max([self.cfl*min([dt_x,dt_y]),self.min_dt])
+
+            # If using an adaptive timestep, slowing increase dt
+            # over the first 20 steps.
+            if self.num_steps <= 20:
+                self.dt *= 1./(5*(21 - self.num_steps))
+        else:
+            self.dt = self.fixed_dt
+
+
+        # Slowly ramp-up the dt for the first 20 steps
+        if self.num_steps <= 20:
+            self.dt *= 1./(5*(21-self.num_steps))
 
     # Initialize the saving
     def initialize_saving(self):
         
-        path = 'Outputs/{0:s}'.format(self.run_name)
-        # If directory already exists, delete it.
-        if os.path.isdir(path):
-           print('Output directory {0:s} already exists. '.format(path) + \
-                 'Warning, deleting everything in the directory.') 
-           shutil.rmtree(path)
-        # Make directory.
-        os.mkdir(path)
+        if not(self.restarting):
+            path = 'Outputs/{0:s}'.format(self.run_name)
+            if not(os.path.isdir('Outputs')):
+                os.mkdir('Outputs')
 
-        if self.animate == 'Save':
-            os.mkdir(path+'/Frames')
+            # If directory already exists, delete it.
+            if os.path.isdir(path):
+               print('Output directory {0:s} already exists. '.format(path) + \
+                     'Warning, deleting everything in the directory.') 
+               shutil.rmtree(path)
+            # Make directory.
+            os.mkdir(path)
+
+            if self.animate == 'Save':
+                os.mkdir(path+'/Frames')
 
         # Initialize saving stuff
         self.save_info()
